A basic test page for the home page of the pcap analyser
This is a basic html page that I worked on in order to create fundamentals that allowed me to work on the backend without having to wait for the real fully completed pages. It houses an upload button with a submit button that only allows you to submit .pcap files to a flask backend.