The current test js script.
This is a test javascript file that will modify elements of both the user page and the results page in order to connect the frontend to the backend. For the home page this will validate that files inputted are in .pcap format and alert the user if they are not. This will then send the pcap file to the flask server for processing, once done this will then reroute to the results page. Upon entering the results page the js will update the DOM and populate the page with results from the flask server that processed the pcap file.